server/scripts/file_upload/requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
psycopg2==2.9.9 